[REFACTOR] Move login into own class
